+++ src/Console/Style/Colors.php
@@ -54,16 +54,16 @@ class Colors
 
         // Check if given foreground color found
         if (isset($this->foreground_colors[$foreground_color])) {
-            $colored_string .= "\033[" . $this->foreground_colors[$foreground_color] . "m";
+            $colored_string .= "\e[" . $this->foreground_colors[$foreground_color] . "m";
         }
-        
+
         // Check if given background color found
         if (isset($this->background_colors[$background_color])) {
-            $colored_string .= "\033[" . $this->background_colors[$background_color] . "m";
+            $colored_string .= "\e[" . $this->background_colors[$background_color] . "m";
         }
 
         // Add string and end coloring
-        $colored_string .= $string . "\033[0m";
+        $colored_string .= $string . "\e[0m";
 
         return $new_line ?  $colored_string . PHP_EOL : $colored_string;
     }